テクスチャちょっとだした

+++ SunnySideBreak/Global.cpp
@@ -1,6 +1,6 @@
 #include "Global.h"
 
-int g_WindowWidth = 1080;
-int g_WindowHeight = 720;
+float g_WindowWidth = 1080.f;
+float g_WindowHeight = 720.f;
 int g_FrameRate = 60;
 const char* g_GameTitle = "SunnySideDownIsDead";
